updating setting screen, preparing delete user function
