+++ lib/main.dart
@@ -3,10 +3,11 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_deadline_management/screens/add_event_screen.dart';
 import 'package:flutter_deadline_management/screens/calendar_screen.dart';
-import 'package:flutter_deadline_management/screens/login_screen.dart';
-import 'package:flutter_deadline_management/screens/registration_screen.dart';
-import 'package:flutter_deadline_management/screens/setting_screen.dart';
-import 'package:flutter_deadline_management/screens/welcome_screen.dart';
+import 'package:flutter_deadline_management/screens/log/login_screen.dart';
+import 'package:flutter_deadline_management/screens/log/registration_screen.dart';
+import 'package:flutter_deadline_management/screens/setting_pages/account.dart';
+import 'package:flutter_deadline_management/screens/setting_pages/setting_screen.dart';
+import 'package:flutter_deadline_management/screens/log/welcome_screen.dart';
 import 'package:flutter_deadline_management/start_up.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/date_symbol_data_local.dart';
@@ -42,6 +43,7 @@ class MyApp extends StatelessWidget {
         SettingScreen.id: (BuildContext context) => SettingScreen(),
         StartUpPage.id: (BuildContext context) => StartUpPage(),
         AddEventScreen.id: (BuildContext context) => AddEventScreen(),
+        SettingAccountScreen.id : (BuildContext context) => SettingAccountScreen(),
       },
 
       // ログインしているかしていないかで最初に表示するページを変える処理
