+++ lib/screens/log/login_screen.dart
@@ -3,43 +3,53 @@ import 'package:flutter/material.dart';
 import 'package:flutter_deadline_management/start_up.dart';
 import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
 
-import '../component/rounded_button.dart';
-import '../constants.dart';
+import '../../component/rounded_button.dart';
+import '../../component/input_decoration_constants.dart';
 
-class RegistrationScreen extends StatefulWidget {
-  static const String id = 'registration';
+class LoginScreen extends StatefulWidget {
+  static const String id = 'login';
   @override
-  _RegistrationScreenState createState() => _RegistrationScreenState();
+  _LoginScreenState createState() => _LoginScreenState();
 }
 
-class _RegistrationScreenState extends State<RegistrationScreen> {
+class _LoginScreenState extends State<LoginScreen> {
   final FirebaseAuth _auth = FirebaseAuth.instance;
   bool isLoad = false;
   String email = '';
   String password = '';
+  // firebaseからのエラーメッセージを表示するための変数
+  String errorMessage = '';
 
-  Future<void> createUserFromEmail() async {
+  Future<void> loginUserFromEmail() async {
     setState(() {
       isLoad = true;
     });
+
     try {
-      final newUser = await _auth.createUserWithEmailAndPassword(
+      // newUserに値を入れる
+      final newUser = await _auth.signInWithEmailAndPassword(
           email: email, password: password);
+
+      // ユーザが確認できたら
       if (newUser != null) {
-        Navigator.pushNamed(context, StartUpPage.id);
+        // 全画面ポップしてスタートあっぷ画面に飛ぶ(スタートアップだけがスタックに存在する
+        Navigator.of(context)
+            .pushNamedAndRemoveUntil(StartUpPage.id, (route) => false);
+
+        print("ログインに成功しました");
       }
       setState(() {
         isLoad = false;
       });
     } on FirebaseAuthException catch (e) {
-      if (e.code == 'email-already-in-use') {
-        print('指定したメールアドレスは登録済みです');
+      if (e.code == 'user-disabled') {
+        errorMessage = 'そのメールアドレスは利用できません';
       } else if (e.code == 'invalid-email') {
-        print('メールアドレスのフォーマットが正しくありません');
-      } else if (e.code == 'operation-not-allowed') {
-        print('指定したメールアドレス・パスワードは現在使用できません');
-      } else if (e.code == 'weak-password') {
-        print('パスワードは６文字以上にしてください');
+        errorMessage = 'メールアドレスのフォーマットが正しくありません';
+      } else if (e.code == 'user-not-found') {
+        errorMessage = 'ユーザーが見つかりません';
+      } else if (e.code == 'wrong-password') {
+        errorMessage = 'パスワードが違います';
       }
     }
     // エラー後にロードを解除する
@@ -68,7 +78,7 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
               children: <Widget>[
                 Flexible(
                   child: Text(
-                    "ユーザー登録",
+                    "ログイン↓↓",
                     style: TextStyle(
                       color: Colors.grey,
                       fontSize: 45.0,
@@ -88,15 +98,18 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
                     email = value;
                   },
                   decoration: kTextFieldDecoration.copyWith(
-                    hintText: 'メールアドレスを入力してください',
+                    hintText: 'メールアドレスを入力して下さい',
                   ),
+
                   // 右下のボタンをNEXTにかえる
                   textInputAction: TextInputAction.next,
+
                   // NEXTを押したらパスワード入力フォームにフォーカスさせる
                   onFieldSubmitted: (_) {
                     FocusScope.of(context)
                         .requestFocus(_passwordFocusNode); // 変更
                   },
+
 //                  バリデーション実装
                   validator: (value) {
                     if (value == null || value.isEmpty) {
@@ -109,7 +122,7 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
                   height: 8.0,
                 ),
 
-                // ここからパスワード用のテキストフィールド
+                // パスワードの入力フォーム
                 TextFormField(
                     obscureText: true,
                     onChanged: (value) {
@@ -119,31 +132,48 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
                     decoration: kTextFieldDecoration.copyWith(
                       hintText: 'パスワードを入力してください',
                     ),
+
                     // パスワード入力フォームに飛ばせるたえのやつ
                     focusNode: _passwordFocusNode,
+
                     //バリデーション実装
                     validator: (value) {
                       if (value == null || value.isEmpty) {
                         return 'パスワードを入力してください';
                       }
                       if (value.length < 6) {
-                        return 'パスワードは6文字以上でないといけません';
+                        return 'パスワードは6文字以上です';
                       }
                       return null;
                     }),
+                SizedBox(
+                  height: 8.0,
+                ),
+
+                // FIREBASEからのエラー表示
+                Center(
+                  child: Text(
+                    errorMessage,
+                    style: TextStyle(
+                      color: Colors.red,
+                      decoration: TextDecoration.underline,
+                      decorationColor: Colors.red,
+                    ),
+                  ),
+                ),
 
                 SizedBox(
                   height: 24.0,
                 ),
 
-                // 登録ボタン
+                // ログインボタン
                 RoundedButton(
                   textColor: Colors.white,
                   color: Colors.blue,
-                  title: '登録',
+                  title: 'ログイン',
                   onPressed: () {
                     if (_formKey.currentState!.validate()) {
-                      createUserFromEmail();
+                      loginUserFromEmail();
                     }
                     print(email);
                     print(password);
